+++ src/Components/GameSettings/index.js
