+++ src/Components/ReadOnlyStrategyChartViewer/index.js
