+++ src/Components/RulesDialog/index.js
