+++ src/Components/ShortcutConfigModal/index.js
