+++ src/Components/StrategyEditor/index.js
